Add QEMU SBSA Patching Support
Adds support to patch QEMU SBSA FW. The actual `QEMU_EFI.fd` file
is 256MB. To reduce the size in this code repo, it is checked in
as an LZMA compressed file which gets down to ~2MB. The script
will automatically decompress it (and any reference image with a
`.lzma` extension) using LZMA decompression. The decompressed
image will automatically be used in the future.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>